+++ src/main/java/no/fintlabs/consumer/config/CaffeineCacheConfig.java
@@ -4,7 +4,7 @@ import com.github.benmanes.caffeine.cache.Cache;
 import com.github.benmanes.caffeine.cache.Caffeine;
 import no.fint.model.resource.Link;
 import no.fintlabs.adapter.models.event.ResponseFintEvent;
-import no.fintlabs.consumer.links.RelationKey;
+import no.fintlabs.consumer.links.relation.RelationKey;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
